Structure adjustment and NextFuction type from express
