+++ graphql-sailboat.d.ts
@@ -1,7 +1,7 @@
 export type SailServerOptions = {
-  path: string | null,
-  graphiql: boolean | null,
-  port: number | null
+  path?: string | null,
+  graphiql?: boolean | null,
+  port?: number | null
 }
 
-export type SailMiddleware = () => never;
+export type SailMiddleware = () => any;
